% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADCD}
\alias{ADCD}
\title{Ferroptosis Drivers}
\format{
A character vector of gene symbols.
}
\source{
FerrDb: http://www.nmdon.cn/fed/
}
\usage{
ADCD
}
\description{
Genes that promote ferroptosis.
}
\keyword{datasets}
