% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Autosis}
\alias{Autosis}
\title{Ferroptosis Suppressors}
\format{
A character vector of gene symbols.
}
\source{
FerrDb: http://www.nmdon.cn/fed/
}
\usage{
Autosis
}
\description{
Genes that suppress ferroptosis.
}
\keyword{datasets}
