% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Mechanism.R
\name{calc_ferro_score}
\alias{calc_ferro_score}
\title{Calculate Ferroptosis (or Custom) Module Scores}
\usage{
calc_ferro_score(seurat_obj, gene_list, name = "FerroScore")
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{gene_list}{A vector of gene symbols (e.g., ferroptosis drivers).}

\item{name}{The name for the metadata column (e.g., "FerroScore").}
}
\value{
A Seurat object with the calculated score added to metadata.
}
\description{
Calculates module scores for a given gene set (e.g., Ferroptosis Drivers).
}
