% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{copykat_data}
\alias{copykat_data}
\title{CopyKAT Prediction Data}
\format{
Data frame with cell IDs and copykat predictions.
}
\source{
CopyKAT analysis
}
\usage{
copykat_data
}
\description{
CopyKAT analysis results for single-cell classification.
}
\keyword{datasets}
