% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cuproptosis_FerrDb}
\alias{Cuproptosis_FerrDb}
\title{Cuproptosis Genes from FerrDb}
\format{
A data frame or list.
}
\source{
FerrDb: http://www.nmdon.cn/fed/
}
\usage{
Cuproptosis_FerrDb
}
\description{
Cuproptosis-related genes from FerrDb database.
}
\keyword{datasets}
