% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Disulfidptosis}
\alias{Disulfidptosis}
\title{Disulfidptosis Genes}
\format{
A data frame or list.
}
\usage{
Disulfidptosis
}
\description{
Genes associated with disulfidptosis cell death.
}
\keyword{datasets}
