% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SCENIC.R
\name{export_scenic_data}
\alias{export_scenic_data}
\title{Export Count Matrix for SCENIC}
\usage{
export_scenic_data(
  seurat_obj,
  output_dir = ".",
  file_name = "scenic_input.csv",
  target_gene = NULL
)
}
\description{
Export Count Matrix for SCENIC
}
