% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ferroptosis_FerrDb}
\alias{Ferroptosis_FerrDb}
\title{Ferroptosis Genes from FerrDb}
\format{
A HCCGeneSet object or data frame.
}
\source{
FerrDb: http://www.nmdon.cn/fed/
}
\usage{
Ferroptosis_FerrDb
}
\description{
Ferroptosis-related genes including drivers, suppressors and markers.
}
\keyword{datasets}
