% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HCC_GeneCards_List}
\alias{HCC_GeneCards_List}
\title{HCC GeneCards Gene List}
\format{
A character vector of gene symbols.
}
\source{
GeneCards: https://www.genecards.org
}
\usage{
HCC_GeneCards_List
}
\description{
Gene list related to hepatocellular carcinoma from GeneCards.
}
\keyword{datasets}
