% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HCCMultiOmics-internal}
\alias{HCCMultiOmics-internal}
\title{Internal Datasets for HCCMultiOmics Package}
\description{
These datasets are used internally by the package functions for
demonstration and analysis purposes.
}
\keyword{datasets}
