% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_CopyKAT.R
\name{identify_malignant_cells}
\alias{identify_malignant_cells}
\title{Identify Malignant Hepatocytes using CopyKAT Data}
\usage{
identify_malignant_cells(
  seurat_obj,
  pred_data = NULL,
  cell_type = "Hepatocytes"
)
}
\description{
Identify Malignant Hepatocytes using CopyKAT Data
}
