% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunogonic_cell_death}
\alias{immunogonic_cell_death}
\title{Immunogenic Cell Death Genes}
\format{
A data frame or list.
}
\usage{
immunogonic_cell_death
}
\description{
Genes associated with immunogenic cell death.
}
\keyword{datasets}
