% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mitotic_death}
\alias{mitotic_death}
\title{Mitotic Death Genes}
\format{
A data frame or list.
}
\usage{
mitotic_death
}
\description{
Genes associated with mitotic death.
}
\keyword{datasets}
