% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parthanatos}
\alias{parthanatos}
\title{Parthanatos Genes}
\format{
A data frame or list.
}
\usage{
parthanatos
}
\description{
Genes associated with parthanatos cell death.
}
\keyword{datasets}
