% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_CopyKAT.R
\name{plot_copykat_stats}
\alias{plot_copykat_stats}
\title{Plot CopyKAT CNV Statistics (Barplots)}
\usage{
plot_copykat_stats(hep_obj, group_col = "site")
}
\description{
Plot CopyKAT CNV Statistics (Barplots)
}
