% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SCENIC.R
\name{plot_earth_density}
\alias{plot_earth_density}
\title{Plot Earth-Tone Density Correlation}
\usage{
plot_earth_density(
  seurat_obj,
  x_feature,
  y_feature,
  x_assay = "RNA",
  y_assay = "AUC"
)
}
\description{
Plot Earth-Tone Density Correlation
}
