% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_Validation.R
\name{plot_gene_diag_score}
\alias{plot_gene_diag_score}
\title{Plot Single Gene Diagnostic Importance (Bar Chart)}
\usage{
plot_gene_diag_score(diag_result, gene = "EZH2")
}
\arguments{
\item{diag_result}{Result object from \code{run_diagnostic_model}.}

\item{gene}{The gene symbol to visualize (e.g., "EZH2").}
}
\value{
A ggplot object.
}
\description{
Visualizes the importance score of a specific gene across RF, SVM, GBM and Average.
}
