% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_lasso_coef}
\alias{plot_lasso_coef}
\title{Plot LASSO Coefficients}
\usage{
plot_lasso_coef(res)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}
}
\description{
Plot LASSO Coefficients
}
