% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_lasso_cv}
\alias{plot_lasso_cv}
\title{Plot LASSO Cross-Validation Curve}
\usage{
plot_lasso_cv(res)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}
}
\description{
Plot LASSO Cross-Validation Curve
}
