% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_model_venn}
\alias{plot_model_venn}
\title{Plot Top Feature Intersection (Venn)}
\usage{
plot_model_venn(res, top_n = 5)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}

\item{top_n}{Number of top features to select from each model for intersection.}
}
\description{
Plot Top Feature Intersection (Venn)
}
