% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Mechanism.R
\name{plot_nature_density}
\alias{plot_nature_density}
\title{绘制 Nature 风格密度散点图 (验证相关性)}
\usage{
plot_nature_density(seurat_obj, x_gene, y_gene)
}
\arguments{
\item{seurat_obj}{Seurat 对象}

\item{x_gene}{X轴基因 (通常是 EZH2)}

\item{y_gene}{Y轴基因 (通常是筛选出来的 Top 1)}
}
\value{
ggplot 对象
}
\description{
绘制 Nature 风格密度散点图 (验证相关性)
}
