% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_rsf_process}
\alias{plot_rsf_process}
\title{Plot RSF Convergence}
\usage{
plot_rsf_process(res)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}
}
\description{
Plot RSF Convergence
}
