% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_rsf_vimp}
\alias{plot_rsf_vimp}
\title{Plot RSF Variable Importance (Lollipop)}
\usage{
plot_rsf_vimp(res)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}
}
\description{
Plot RSF Variable Importance (Lollipop)
}
