% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SingleCell.R
\name{plot_sc_gene}
\alias{plot_sc_gene}
\title{Plot Target Gene Expression}
\usage{
plot_sc_gene(
  seurat_obj,
  gene = "EZH2",
  color_low = "#F0F0F0",
  color_high = "#B48283"
)
}
\description{
Plot Target Gene Expression
}
\note{
FeaturePlot 配色改为：极浅灰(#F0F0F0) -> 低饱和红褐(#B48283)
}
