% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SingleCell.R
\name{plot_sc_gene_by_site}
\alias{plot_sc_gene_by_site}
\title{Plot Gene Statistics by Site in Specific Cell Type}
\usage{
plot_sc_gene_by_site(
  seurat_obj,
  gene = "EZH2",
  cell_type = "Hepatocytes",
  group_col = "site",
  plot_type = 2
)
}
\arguments{
\item{plot_type}{Integer. 1 for Average Expression, 2 for Proportion (\%). Default is 2.}
}
\description{
Plot Gene Statistics by Site in Specific Cell Type
}
