% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SingleCell.R
\name{plot_sc_gene_stat}
\alias{plot_sc_gene_stat}
\title{Plot Gene Statistics per Cell Type}
\usage{
plot_sc_gene_stat(seurat_obj, gene = "EZH2", plot_type = 2)
}
\description{
Plot Gene Statistics per Cell Type
}
