% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Mechanism.R
\name{plot_tradeseq_trends}
\alias{plot_tradeseq_trends}
\title{Plot Smooth Trends}
\usage{
plot_tradeseq_trends(sce_obj, plot_genes, colors = NULL)
}
\arguments{
\item{sce_obj}{The fitted SCE object.}

\item{plot_genes}{Genes to plot.}

\item{colors}{Named vector of colors.}
}
\description{
Plot Smooth Trends
}
