% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{plot_xgb_imp}
\alias{plot_xgb_imp}
\title{Plot XGBoost Importance}
\usage{
plot_xgb_imp(res)
}
\arguments{
\item{res}{Object of class HCCPrognosticResult}
}
\description{
Plot XGBoost Importance
}
