% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_GeneSet_Manager.R
\name{print.HCCGeneSet}
\alias{print.HCCGeneSet}
\title{Print method for HCCGeneSet objects}
\usage{
\method{print}{HCCGeneSet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{HCCGeneSet}.}

\item{...}{Additional arguments.}
}
\description{
Custom print function to display gene set information neatly.
}
