% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Trajectory.R
\name{reprocess_malignant_cells}
\alias{reprocess_malignant_cells}
\title{Reprocess Malignant Cells (Harmony Integration)}
\usage{
reprocess_malignant_cells(
  seurat_obj,
  batch_col = "patient",
  theta = 3,
  resolution = 0.5
)
}
\description{
Reprocess Malignant Cells (Harmony Integration)
}
