% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_Bulk_Discovery.R
\name{run_intersection}
\alias{run_intersection}
\title{Perform Intersection Analysis for Target Discovery}
\usage{
run_intersection(gene_set_obj, degs_list = NULL, hcc_list = NULL)
}
\arguments{
\item{gene_set_obj}{An object of class \code{HCCGeneSet}.}

\item{degs_list}{Vector of strings. Background DEGs (defaults to internal data).}

\item{hcc_list}{Vector of strings. Background HCC genes (defaults to internal data).}
}
\value{
A list containing:
\item{candidates}{Vector of overlapping gene symbols (The Center intersection)}
\item{venn_lists}{A named list of the three gene sets used for plotting}
\item{counts}{A list of counts for each region}
}
\description{
This function intersects the user-provided gene set with internal TCGA-LIHC DEGs
and GeneCards HCC-related genes to prioritize candidates.
}
