% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SCENIC.R
\name{run_mediation_analysis}
\alias{run_mediation_analysis}
\title{Run Mediation Analysis with Bootstrap}
\usage{
run_mediation_analysis(
  seurat_obj,
  target_gene,
  mediator_tf,
  downstream_gene,
  n_sims = 500
)
}
\description{
Run Mediation Analysis with Bootstrap
}
