% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleA_MachineLearning.R
\name{run_prognostic_model}
\alias{run_prognostic_model}
\title{Run Ultimate Rigorous Prognostic Screening}
\usage{
run_prognostic_model(candidates, expr_mat = NULL, clin_data = NULL, seed = 123)
}
\arguments{
\item{candidates}{Vector of gene symbols.}

\item{expr_mat}{Expression matrix.}

\item{clin_data}{Clinical data frame.}

\item{seed}{Random seed for reproducibility (Default: 123). Change this if result varies!}
}
\value{
An object of class \code{HCCPrognosticResult}.
}
\description{
Run Ultimate Rigorous Prognostic Screening
}
