% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SCENIC.R
\name{run_tf_enrichment_analysis}
\alias{run_tf_enrichment_analysis}
\title{Run TF Enrichment Analysis (Lollipop + Heatmap Linkage)}
\usage{
run_tf_enrichment_analysis(seurat_obj, target_gene = "EZH2")
}
\arguments{
\item{seurat_obj}{Seurat object (Raw object, function will filter target_gene > 0 internally).}

\item{target_gene}{Gene to correlate with (default "EZH2").}
}
\value{
A list containing \code{lollipop_plot} and \code{heatmap_obj}.
}
\description{
Run TF Enrichment Analysis (Lollipop + Heatmap Linkage)
}
