% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Trajectory.R
\name{run_trajectory_analysis}
\alias{run_trajectory_analysis}
\title{Run Slingshot Trajectory Analysis}
\usage{
run_trajectory_analysis(seurat_obj, start_cluster = "1")
}
\description{
Run Slingshot Trajectory Analysis
}
