% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_SCENIC.R
\name{screen_mediator_tfs}
\alias{screen_mediator_tfs}
\title{Screen for Mediator Transcription Factors}
\usage{
screen_mediator_tfs(
  seurat_obj,
  target_gene = "EZH2",
  downstream_gene = "SLC7A11"
)
}
\arguments{
\item{seurat_obj}{Seurat object with "AUC" assay.}

\item{target_gene}{The upstream driver gene (e.g., "EZH2").}

\item{downstream_gene}{The downstream effector gene (e.g., "SLC7A11").}
}
\value{
A list containing the results data frame (\code{res_df}) and a screening plot (\code{plot}).
}
\description{
Screen for Mediator Transcription Factors
}
