% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleB_Trajectory.R
\name{subset_trajectory}
\alias{subset_trajectory}
\title{Subset Cells on a Specific Trajectory}
\usage{
subset_trajectory(seurat_obj, sds, curve_id = 1)
}
\description{
Subset Cells on a Specific Trajectory
}
