% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_Clin_Data}
\alias{TCGA_Clin_Data}
\title{TCGA Liver Cancer Clinical Data}
\format{
A data frame with columns: SampleID, Group, OS, OS.time, ...
}
\source{
TCGA Research Network: https://www.cancer.gov/tcga
}
\usage{
TCGA_Clin_Data
}
\description{
Clinical information for TCGA-LIHC samples including survival data.
}
\keyword{datasets}
