% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_Expr_Mat}
\alias{TCGA_Expr_Mat}
\title{TCGA Liver Cancer Expression Matrix}
\format{
A matrix with genes as rows and samples as columns.
}
\source{
TCGA Research Network: https://www.cancer.gov/tcga
}
\usage{
TCGA_Expr_Mat
}
\description{
Gene expression matrix for TCGA-LIHC samples (normalized).
}
\keyword{datasets}
