% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_LIHC_DEGs}
\alias{TCGA_LIHC_DEGs}
\title{TCGA-LIHC Differentially Expressed Genes}
\format{
A character vector of gene symbols.
}
\source{
TCGA Research Network: https://www.cancer.gov/tcga
}
\usage{
TCGA_LIHC_DEGs
}
\description{
List of differentially expressed genes identified in TCGA-LIHC.
}
\keyword{datasets}
